requirements.txt
google-generativeai
discord
